+++ docs/CHANGELOG.md
@@ -7,6 +7,8 @@
 - feat: support setting global testcase timeout and step timeout
 - change: set http request timeout default to 120s
 - fix: insert response cookies into request for redirect requests
+- fix: support log debug level for load testing
+- fix: failed to load json/data content in api reference
 
 ## v4.1.4 (2022-06-17)