+++ docs/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## v4.1.5 (2022-06-26)
+## v4.1.5 (2022-06-27)
 
 **go version**